+++ battleship/nbproject/project.properties
@@ -53,7 +53,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=battleship.Battleship
+main.class=citbyui260.section03.battleship.game.Battleship
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
