+++ battleship/src/citbyui260/section03/battleship/game/Player.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package battleship;
+package citbyui260.section03.battleship.game;
 import citbyui260.section03.battleship.ships.Boat;
 import citbyui260.section03.battleship.boards.ShotBoard;
 import citbyui260.section03.battleship.boards.ShipBoard;
